Refactored ProfileController with defensive programming

+++ VideoAppBLL/Service/VideoService.cs
@@ -2,8 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using VideoAppBLL.BusinessObjects;
-using VideoAppDAL.Entities;
+using VideoAppBLL.Converters;
+using VideoAppBLL.Interfaces;
 using VideoAppDAL.Interfaces;
+using static VideoAppBLL.Converters.VideoConverter;
 
 namespace VideoAppBLL.Service
 {
@@ -31,7 +33,7 @@ namespace VideoAppBLL.Service
             throw new NotImplementedException();
         }
 
-        public IEnumerable<VideoBO> GetAll()
+        public IList<VideoBO> GetAll()
         {
             using (var unitOfWork = _facade.UnitOfWork)
             {
@@ -44,7 +46,8 @@ namespace VideoAppBLL.Service
         {
             using (var unitOfWork = _facade.UnitOfWork)
             {
-                return Convert(unitOfWork.VideoRepository.GetById(id));
+                var videoFromDB = unitOfWork.VideoRepository.GetById(id);
+                return videoFromDB == null ? null : Convert(videoFromDB);
             }
         }
 
@@ -52,6 +55,8 @@ namespace VideoAppBLL.Service
         {
             using (var unitOfWork = _facade.UnitOfWork)
             {
+                var videoFromDB = unitOfWork.VideoRepository.GetById(id);
+                if (videoFromDB == null) return false;
                 var videoDeleted = unitOfWork.VideoRepository.Delete(id);
                 unitOfWork.Complete();
                 return videoDeleted;
@@ -67,7 +72,7 @@ namespace VideoAppBLL.Service
                 if (videoFromRepo == null) return null;
 
                 videoFromRepo.Title = entityToUpdate.Title;
-                videoFromRepo.Genre = entityToUpdate.Genre;
+                videoFromRepo.Genre = GenreConverter.Convert(entityToUpdate.Genre);
                 var updatedVideo = unitOfWork.VideoRepository.Update(videoFromRepo);
                 unitOfWork.Complete();
                 return Convert(updatedVideo);
@@ -81,36 +86,6 @@ namespace VideoAppBLL.Service
                 unitOfWork.VideoRepository.ClearAll();
                 unitOfWork.Complete();
             }
-        }
-
-        /// <summary>
-        /// Convert Video to VideoBO
-        /// </summary>
-        /// <param name="video"></param>
-        /// <returns>VideoBO</returns>
-        private VideoBO Convert(Video video)
-        {
-            return new VideoBO()
-            {
-                Id = video.Id,
-                Title = video.Title,
-                Genre = video.Genre
-            };
-        }
-
-        /// <summary>
-        /// Convert VideoBO to VIDEO
-        /// </summary>
-        /// <param name="video"></param>
-        /// <returns>Video</returns>
-        private Video Convert(VideoBO video)
-        {
-            return new Video()
-            {
-                Id = video.Id,
-                Title = video.Title,
-                Genre = video.Genre
-            };
         }
     }
 }
